Update SDI12_boards.cpp
I found another bug for your lib using the ATtiny(85).

You set the prescaler for the timer on the register TCCR1A (chapter 13 in Attiny complete datasheet). This only works, when your Attiny is in the "ATtiny15 compatible mode" (can be set with fuses). If your Tiny is not in that mode (and I think it usually isnt) the prescaler is then wrong and the Bits are way to fast.
Instead you need to write to the TCCR1 register (chapter 12 in Attiny complete datasheet). Then the 1200 baud rate is right and the sensor is responding. (Tested with Attiny85@8Mhz)

With best regards